requirements.txt
Flask==2.3.3
google-cloud-firestore==2.11.1
Flask-WTF==1.1.1
python-dotenv==1.0.0gunicorn
